+++ src/solution_f8ff0b80.py
@@ -1,5 +1,5 @@
 import numpy as np
-from process_print import process
+from process_data import process
 import json
